requirements.txt
django == 1.9.5
paramiko == 1.16.0
django-extensions == 1.6.1
django-rest-swagger == 0.3.6
djangorestframework == 3.3.3
django-cors-headers == 1.1.0
MySQL-python == 1.2.5
gunicorn == 19.4.5
django-filter == 0.13.0
celery == 3.1.23
django-celery == 3.1.17
redis == 2.10.5
httplib2 == 0.9.2
flower == 0.9.1
python-memcached == 1.57